tree2: remove editable-tree-1 (#18169)
## Description

Remove editable-tree-1 and APIs related to it.

## Breaking Changes

Users of editable-tree-1 must migrate to editable-tree-2.
