+++ .changeset/rude-tips-visit.md
@@ -1,0 +1,7 @@
+---
+"@fluid-experimental/tree2": minor
+---
+
+Remove editable-tree-1
+
+Remove editable-tree-1 and APIs related to it. Users must migrate to editable-tree-2.
