+++ experimental/dds/tree2/src/feature-libraries/editable-tree-2/editableTreeTypes.ts
@@ -75,7 +75,7 @@ export interface Tree<out TSchema = unknown> {
 }
 
 /**
- * Status of the tree that a particular node in {@link EditableTree} and {@link UntypedTree} belongs to.
+ * Status of the tree that a particular node in {@link Tree} belongs to.
  * @alpha
  */
 export enum TreeStatus {
