+++ experimental/dds/tree2/src/feature-libraries/editable-tree-2/nodeKeys.ts
@@ -9,7 +9,7 @@ import { ObjectNode } from "./editableTreeTypes";
 /**
  * A collection of utilities for managing {@link StableNodeKey}s.
  * A node key can be assigned to a node and allows that node to be easily retrieved from the tree at a later time. (see `nodeKey.map`).
- * @remarks {@link LocalNodeKey}s are put on nodes via a special field (see {@link localNodeKeySymbol}.
+ * @remarks {@link LocalNodeKey}s are put on {@link ObjectNode}s via a special field.
  * A node with a node key in its schema must always have a node key.
  * @alpha
  */
