--- experimental/dds/tree2/src/feature-libraries/editable-tree/fence.json
@@ -1,6 +1,0 @@
-{
-	"tags": ["forest-proxy"],
-	"exports": ["index"],
-	// TODO: "feature-libraries" dependency is for defaultFieldKinds, and should be removed once proper view schema are provided.
-	"imports": ["util", "events", "core", "feature-libraries", "modular-schema", "node-key"]
-}
