--- experimental/dds/tree2/src/feature-libraries/editable-tree/index.ts
@@ -1,61 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export {
-	EditableTree,
-	EditableField,
-	EditableTreeOrPrimitive,
-	proxyTargetSymbol,
-	UnwrappedEditableTree,
-	UnwrappedEditableField,
-	areCursors,
-	localNodeKeySymbol,
-	setField,
-} from "./editableTreeTypes";
-
-export { isEditableField } from "./editableField";
-export { isEditableTree } from "./editableTree";
-export {
-	createDataBinderBuffering,
-	createDataBinderDirect,
-	createDataBinderInvalidating,
-	createBinderOptions,
-	createFlushableBinderOptions,
-	DataBinder,
-	BinderOptions,
-	Flushable,
-	FlushableBinderOptions,
-	FlushableDataBinder,
-	MatchPolicy,
-	SubtreePolicy,
-	BindSyntaxTree,
-	indexSymbol,
-	BindPolicy,
-	BindTree,
-	BindTreeDefault,
-	DownPath,
-	BindPath,
-	PathStep,
-	BindingType,
-	BindingContextType,
-	VisitorBindingContext,
-	BindingContext,
-	DeleteBindingContext,
-	InsertBindingContext,
-	BatchBindingContext,
-	InvalidationBindingContext,
-	OperationBinderEvents,
-	InvalidationBinderEvents,
-	CompareFunction,
-	BinderEventsCompare,
-	AnchorsCompare,
-	toDownPath,
-	comparePipeline,
-	compileSyntaxTree,
-} from "./editableTreeBinder";
-
-export { EditableTreeContext, getEditableTreeContext } from "./editableTreeContext";
-
-export { isPrimitive, treeStatusFromPath } from "./utilities";
