+++ experimental/dds/tree2/src/feature-libraries/schema-aware/schemaAware.ts
@@ -3,17 +3,8 @@
  * Licensed under the MIT License.
  */
 
-import { assert } from "@fluidframework/core-utils";
-import { PrimitiveValueSchema, TreeNodeSchemaIdentifier, TreeValue, ValueSchema } from "../../core";
-import {
-	ContextuallyTypedNodeData,
-	FluidSerializableReadOnly,
-	MarkedArrayLike,
-	PrimitiveValue,
-	isFluidHandle,
-	typeNameSymbol,
-	valueSymbol,
-} from "../contextuallyTyped";
+import { TreeNodeSchemaIdentifier, TreeValue, ValueSchema } from "../../core";
+import { ContextuallyTypedNodeData, typeNameSymbol, valueSymbol } from "../contextuallyTyped";
 import { Multiplicity } from "../modular-schema";
 import {
 	InternalTypedSchemaTypes,
@@ -21,15 +12,7 @@ import {
 	TreeNodeSchema,
 	AllowedTypes,
 } from "../typed-schema";
-import {
-	UntypedField,
-	UntypedTree,
-	UntypedTreeCore,
-	contextSymbol,
-	typeSymbol,
-} from "../untypedTree";
 import { Assume, FlattenKeys, _InlineTrick } from "../../util";
-import { UntypedOptionalField, UntypedSequenceField, UntypedValueField } from "./partlyTyped";
 import { TypedValueOrUndefined } from "./schemaAwareUtil";
 
 /**
@@ -60,28 +43,6 @@ export const enum ApiMode {
 	 */
 	Flexible,
 	/**
-	 * Similar to what EditableTree uses.
-	 * No flexibility in representation.
-	 * Fields are unwrapped (see `EditableUnwrapped`).
-	 *
-	 * TODO: fix ways this differs from editable tree:
-	 * - Does not do primary field inlining.
-	 * - Primitive node handling might not match.
-	 */
-	Editable,
-	/**
-	 * Editable, but primitive nodes are unwrapped to the primitive values.
-	 */
-	EditableUnwrapped,
-	/**
-	 * Always use full node objects for everything.
-	 *
-	 * Fields are still shaped based on their multiplicity.
-	 *
-	 * TODO: test and fix
-	 */
-	Wrapped,
-	/**
 	 * Simplified version of Flexible.
 	 *
 	 * Primitive values are always unwrapped.
@@ -102,22 +63,6 @@ export type CollectOptions<
 	[ApiMode.Flexible]: EmptyObject extends TTypedFields
 		? TypedValueOrUndefined<TValueSchema> | FlexibleObject<TValueSchema, TName>
 		: FlexibleObject<TValueSchema, TName> & TTypedFields;
-	[ApiMode.Editable]: {
-		[typeNameSymbol]: TName & TreeNodeSchemaIdentifier;
-	} & ValuePropertyFromSchema<TValueSchema> &
-		TTypedFields &
-		UntypedTreeCore;
-	[ApiMode.EditableUnwrapped]: [EmptyObject, TValueSchema] extends [
-		TTypedFields,
-		PrimitiveValueSchema,
-	]
-		? TypedValueOrUndefined<TValueSchema>
-		: // TODO: primary field unwrapping
-		  CollectOptions<ApiMode.Editable, TTypedFields, TValueSchema, TName>;
-	[ApiMode.Wrapped]: {
-		[typeNameSymbol]: TName;
-		[valueSymbol]: TypedValueOrUndefined<TValueSchema>;
-	} & TTypedFields;
 	[ApiMode.Simple]: EmptyObject extends TTypedFields
 		? TypedValueOrUndefined<TValueSchema>
 		: FlexibleObject<TValueSchema, TName> & TTypedFields;
@@ -153,10 +98,7 @@ export type TypedFields<
 > = [
 	TFields extends { [key: string]: TreeFieldSchema }
 		? {
-				-readonly [key in keyof TFields]: TypedField<
-					TFields[key],
-					Mode extends ApiMode.Editable ? ApiMode.EditableUnwrapped : Mode
-				>;
+				-readonly [key in keyof TFields]: TypedField<TFields[key], Mode>;
 		  }
 		: EmptyObject,
 ][_InlineTrick];
@@ -165,7 +107,7 @@ export type TypedFields<
  * `TreeFieldSchema` to `TypedField`. May unwrap to child depending on Mode and FieldKind.
  * @alpha
  */
-export type TypedField<TField extends TreeFieldSchema, Mode extends ApiMode = ApiMode.Editable> = [
+export type TypedField<TField extends TreeFieldSchema, Mode extends ApiMode> = [
 	ApplyMultiplicity<
 		TField["kind"]["multiplicity"],
 		AllowedTypesToTypedTrees<Mode, TField["allowedTypes"]>,
@@ -180,44 +122,13 @@ export type TypedField<TField extends TreeFieldSchema, Mode extends ApiMode = Ap
 export type ApplyMultiplicity<
 	TMultiplicity extends Multiplicity,
 	TypedChild,
-	Mode extends ApiMode,
+	_Mode extends ApiMode,
 > = {
 	[Multiplicity.Forbidden]: undefined;
-	[Multiplicity.Optional]: Mode extends ApiMode.Editable
-		? EditableOptionalField<TypedChild>
-		: undefined | TypedChild;
-	[Multiplicity.Sequence]: Mode extends ApiMode.Editable | ApiMode.EditableUnwrapped
-		? EditableSequenceField<TypedChild>
-		: TypedChild[];
-	[Multiplicity.Single]: Mode extends ApiMode.Editable
-		? EditableValueField<TypedChild>
-		: TypedChild;
+	[Multiplicity.Optional]: undefined | TypedChild;
+	[Multiplicity.Sequence]: TypedChild[];
+	[Multiplicity.Single]: TypedChild;
 }[TMultiplicity];
-
-// TODO: add strong typed `getNode`.
-export type EditableField<TypedChild> = UntypedField & MarkedArrayLike<TypedChild>;
-
-// TODO: add strong typed `getNode`.
-/**
- * @alpha
- */
-export type EditableSequenceField<TypedChild> = [
-	UntypedSequenceField & MarkedArrayLike<TypedChild>,
-][_InlineTrick];
-
-/**
- * @alpha
- */
-export type EditableValueField<TypedChild> = [
-	UntypedValueField & MarkedArrayLike<TypedChild>,
-][_InlineTrick];
-
-/**
- * @alpha
- */
-export type EditableOptionalField<TypedChild> = [
-	UntypedOptionalField & MarkedArrayLike<TypedChild>,
-][_InlineTrick];
 
 /**
  * Takes in `AllowedTypes` and returns a TypedTree union.
@@ -256,27 +167,18 @@ export type TypeArrayToTypedTreeArray<Mode extends ApiMode, T extends readonly T
  * @alpha
  */
 export type UntypedApi<Mode extends ApiMode> = {
-	[ApiMode.Editable]: UntypedTree;
-	[ApiMode.EditableUnwrapped]: UntypedTree | PrimitiveValue;
 	[ApiMode.Flexible]: ContextuallyTypedNodeData;
 	[ApiMode.Simple]: ContextuallyTypedNodeData;
-	[ApiMode.Wrapped]: UntypedTree;
 }[Mode];
 
 /**
  * Generate a schema aware API for a single tree schema.
  * @alpha
  */
-export type TypedNode<
-	TSchema extends TreeNodeSchema,
-	Mode extends ApiMode = ApiMode.Editable,
-> = FlattenKeys<
+export type TypedNode<TSchema extends TreeNodeSchema, Mode extends ApiMode> = FlattenKeys<
 	CollectOptions<
 		Mode,
-		TypedFields<
-			Mode extends ApiMode.Editable ? ApiMode.EditableUnwrapped : Mode,
-			TSchema["objectNodeFieldsObject"]
-		>,
+		TypedFields<Mode, TSchema["objectNodeFieldsObject"]>,
 		TSchema["leafValue"],
 		TSchema["name"]
 	>
@@ -291,33 +193,3 @@ export type NodeDataFor<Mode extends ApiMode, TSchema extends TreeNodeSchema> = 
 	TSchema,
 	Mode
 >;
-
-/**
- * Check if an `UntypedTreeCore` has a specific schema, and if it does, cast it to use `ApiMode.Editable` with that schema.
- * Provided schema must be included in the schema for the tree being viewed (getting this wrong will error).
- * @alpha
- */
-export function downCast<TSchema extends TreeNodeSchema>(
-	schema: TSchema,
-	tree: UntypedTreeCore,
-): tree is TypedNode<TSchema> {
-	assert(typeof tree === "object", 0x72b /* downCast only valid on wrapped nodes */);
-	assert(tree !== null, 0x7d5 /* downCast only valid on wrapped nodes */);
-	assert(
-		!isFluidHandle(tree as unknown as FluidSerializableReadOnly),
-		0x7d6 /* downCast only valid on wrapped nodes */,
-	);
-
-	const contextSchema = tree[contextSymbol].schema;
-	const lookedUp = contextSchema.nodeSchema.get(schema.name);
-	// TODO: for this to pass, schematized view must have the view schema, not just stored schema.
-	assert(lookedUp === schema, 0x68c /* cannot downcast to a schema the tree is not using */);
-
-	// TODO: make this actually work
-	const matches = tree[typeSymbol] === schema;
-	assert(
-		matches === (tree[typeSymbol].name === schema.name),
-		0x68d /* schema object identity comparison should match identifier comparison */,
-	);
-	return matches;
-}
