+++ experimental/dds/tree2/src/index.ts
@@ -137,22 +137,9 @@ export {
 	Multiplicity,
 	isNeverField,
 	FullSchemaPolicy,
-	UnwrappedEditableField,
-	isEditableTree,
-	isEditableField,
-	EditableTreeContext,
-	UnwrappedEditableTree,
-	EditableTreeOrPrimitive,
-	EditableTree,
-	EditableField,
-	isPrimitive,
 	getPrimaryField,
-	typeSymbol,
 	typeNameSymbol,
 	valueSymbol,
-	proxyTargetSymbol,
-	getField,
-	contextSymbol,
 	ContextuallyTypedNodeDataObject,
 	ContextuallyTypedNodeData,
 	MarkedArrayLike,
@@ -163,7 +150,6 @@ export {
 	StableNodeKey,
 	LocalNodeKey,
 	compareLocalNodeKeys,
-	localNodeKeySymbol,
 	IDefaultEditBuilder,
 	ValueFieldEditBuilder,
 	OptionalFieldEditBuilder,
@@ -174,22 +160,13 @@ export {
 	singleStackTreeCursor,
 	CursorAdapter,
 	CursorWithNode,
-	parentField,
 	EditableTreeEvents,
-	on,
 	InternalTypedSchemaTypes,
 	SchemaAware,
 	ArrayLikeMut,
 	FieldKinds,
 	ContextuallyTypedFieldData,
 	cursorFromContextualData,
-	UntypedField,
-	UntypedTree,
-	UntypedTreeContext,
-	UntypedTreeCore,
-	UnwrappedUntypedField,
-	UnwrappedUntypedTree,
-	UntypedTreeOrPrimitive,
 	AllowedTypes,
 	TreeNodeSchema,
 	TreeSchema,
@@ -203,46 +180,9 @@ export {
 	LazyTreeNodeSchema,
 	FieldGenerator,
 	TreeDataContext,
-	createDataBinderBuffering,
-	createDataBinderDirect,
-	createDataBinderInvalidating,
-	createBinderOptions,
-	createFlushableBinderOptions,
-	DataBinder,
-	BinderOptions,
-	Flushable,
-	FlushableBinderOptions,
-	FlushableDataBinder,
-	MatchPolicy,
-	SubtreePolicy,
-	BindSyntaxTree,
-	indexSymbol,
-	BindPolicy,
-	BindTree,
-	BindTreeDefault,
-	DownPath,
-	BindPath,
-	PathStep,
-	BindingType,
-	BindingContextType,
-	BindingContext,
-	VisitorBindingContext,
-	DeleteBindingContext,
-	InsertBindingContext,
-	BatchBindingContext,
-	InvalidationBindingContext,
-	OperationBinderEvents,
-	InvalidationBinderEvents,
-	CompareFunction,
-	BinderEventsCompare,
-	AnchorsCompare,
-	toDownPath,
-	comparePipeline,
-	compileSyntaxTree,
 	nodeKeyFieldKey,
 	SchemaLintConfiguration,
 	TreeStatus,
-	treeStatus,
 	FieldNode,
 	FlexibleFieldContent,
 	FlexibleNodeContent,
