+++ experimental/dds/tree2/src/shared-tree/sharedTreeView.ts
@@ -23,15 +23,11 @@ import {
 } from "../core";
 import { HasListeners, IEmitter, ISubscribable, createEmitter } from "../events";
 import {
-	UnwrappedEditableField,
-	EditableTreeContext,
 	IDefaultEditBuilder,
 	DefaultChangeset,
 	buildForest,
 	DefaultChangeFamily,
-	getEditableTreeContext,
 	DefaultEditBuilder,
-	NewFieldContent,
 	NodeKeyManager,
 	TreeFieldSchema,
 	TreeSchema,
@@ -82,37 +78,6 @@ export interface ViewEvents {
  * @alpha
  */
 export interface ISharedTreeView extends AnchorLocator {
-	/**
-	 * Gets the root field of the tree.
-	 *
-	 * See {@link EditableTreeContext.unwrappedRoot} on how its setter works.
-	 *
-	 * Currently this editable tree's fields do not update on edits,
-	 * so holding onto this root object across edits will only work if it's an unwrapped node.
-	 * TODO: Fix this issue.
-	 *
-	 * Currently any access to this view of the tree may allocate cursors and thus require
-	 * `context.prepareForEdit()` before editing can occur.
-	 *
-	 * @deprecated Use {@link ISharedTreeView2} and editable tree 2.
-	 */
-	// TODO: either rename this or `EditableTreeContext.unwrappedRoot` to avoid name confusion.
-	get root(): UnwrappedEditableField;
-
-	/**
-	 * Sets the content of the root field of the tree.
-	 *
-	 * See {@link EditableTreeContext.unwrappedRoot} on how this works.
-	 * @deprecated Use {@link ISharedTreeView2} and editable tree 2.
-	 */
-	setContent(data: NewFieldContent): void;
-
-	/**
-	 * Context for controlling the EditableTree-1 nodes.
-	 * @deprecated Use {@link ISharedTreeView2} and editable tree 2.
-	 */
-	readonly context: EditableTreeContext;
-
 	/**
 	 * Read and Write access for schema stored in the document.
 	 *
@@ -232,7 +197,6 @@ export function createSharedTreeView(args?: {
 			},
 			changeFamily,
 		);
-	const context = getEditableTreeContext(forest, schema, branch.editor);
 	const events = args?.events ?? createEmitter();
 
 	const transaction = new Transaction(branch);
@@ -243,7 +207,6 @@ export function createSharedTreeView(args?: {
 		changeFamily,
 		schema,
 		forest,
-		context,
 		events,
 		args?.removedTrees,
 	);
@@ -330,7 +293,6 @@ export class SharedTreeView implements ISharedTreeBranchView {
 		private readonly changeFamily: DefaultChangeFamily,
 		public readonly storedSchema: StoredSchemaRepository,
 		public readonly forest: IEditableForest,
-		public readonly context: EditableTreeContext,
 		public readonly events: ISubscribable<ViewEvents> &
 			IEmitter<ViewEvents> &
 			HasListeners<ViewEvents>,
@@ -412,7 +374,6 @@ export class SharedTreeView implements ISharedTreeBranchView {
 		const storedSchema = new InMemoryStoredSchemaRepository(this.storedSchema);
 		const forest = this.forest.clone(storedSchema, anchors);
 		const branch = this.branch.fork();
-		const context = getEditableTreeContext(forest, storedSchema, branch.editor);
 		const transaction = new Transaction(branch);
 		return new SharedTreeView(
 			transaction,
@@ -420,7 +381,6 @@ export class SharedTreeView implements ISharedTreeBranchView {
 			this.changeFamily,
 			storedSchema,
 			forest,
-			context,
 			createEmitter(),
 			this.removedTrees.clone(),
 		);
@@ -448,14 +408,6 @@ export class SharedTreeView implements ISharedTreeBranchView {
 		if (disposeView) {
 			view.dispose();
 		}
-	}
-
-	public get root(): UnwrappedEditableField {
-		return this.context.unwrappedRoot;
-	}
-
-	public setContent(data: NewFieldContent) {
-		this.context.setContent(data);
 	}
 
 	/**
