+++ experimental/dds/tree2/src/test/feature-libraries/schema-aware/schemaComplex.ts
@@ -36,9 +36,6 @@ export const rootFieldSchema = SchemaBuilder.required([stringTaskSchema, listTas
 export const appSchemaData = builder.intoSchema(rootFieldSchema);
 
 // Schema aware types
-export type StringTask = SchemaAware.TypedNode<typeof stringTaskSchema>;
-
-export type ListTask = SchemaAware.TypedNode<typeof listTaskSchema>;
 
 type FlexibleListTask = SchemaAware.TypedNode<typeof listTaskSchema, SchemaAware.ApiMode.Flexible>;
